+++ tiger-testenv-mgr/src/frontend/src/components/global_configuration/ConfigurationValueCell.vue
@@ -82,4 +82,8 @@ function clickExpandButton() {
 .multi-line {
   white-space: normal;
 }
+
+.row > * {
+  margin-bottom: var(--bs-gutter-y);
+}
 </style>